Created a Book object, renamed original book.java to main.java
I created a book object to store data relevant to each book in the bookstore. So far it only has parameters relevant to my end of the project but it can be modified to include more such as the author as well.

Creating a book object caused conflicts with the original java file with the SQL implementation, so for now I renamed it to "main"