FIXED POST AND GET FOR REVIEW
THIS TOOK LONGER THAN EXPECTED. Because for some reason Review kept trying to get the internal ID for the book instead of the ISBN. Because the ISBN is linked to the interal id, but the user won't know that.  So I had to force it to just look for ISBN and it finally worked! Hoping the next one isn't going to take as long