Implemented Final Feature for Wishlist
Added in the Delete option. It deletes the book from the wishlist and adds it to the user's shoppingcart. Make sure to add a different book first to the wishlist as you can't tell the difference. If we had a extra person do the shopping cart feature, we may be able to tell how many books a person has, but we currently don't.